# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OpenApi documentation for Tree Nodes",
        "description": "Swagger OpenApi documentation for Tree Nodes",
        "contact": {
            "email": "info@website.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.1",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
        "/api/nodes": {
            "get": {
                "tags": [
                    "nodes"
                ],
                "summary": "All nodes paginated",
                "description": "Display a listing of the resource.\nShowing all nodes paginated.",
                "operationId": "6a49c5145f15e738bcec4f33191ed0c4",
                "parameters": [
                    {
                        "name": "X-Header-Timezone",
                        "in": "header",
                        "description": "Get date fields in specific Timezone",
                        "required": false,
                        "examples": {
                            "Caracas": {
                                "summary": "Caracas",
                                "value": "America/Caracas"
                            },
                            "Bogota": {
                                "summary": "Bogota",
                                "value": "America/Bogota"
                            },
                            "Los Angeles": {
                                "summary": "Los Angeles",
                                "value": "America/Los_Angeles"
                            }
                        }
                    },
                    {
                        "name": "X-Header-Lang",
                        "in": "header",
                        "description": "Get title field in specific language",
                        "required": false,
                        "examples": {
                            "en": {
                                "summary": "English",
                                "value": "en"
                            },
                            "es": {
                                "summary": "Spanish",
                                "value": "es"
                            },
                            "fr": {
                                "summary": "French",
                                "value": "fr"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "All nodes paginated."
                    },
                    "500": {
                        "description": "An error has happened."
                    }
                }
            },
            "post": {
                "tags": [
                    "nodes"
                ],
                "summary": "Storing a new node",
                "description": "Store a new created node in storage.\nStoring a new node in a tree or creating root node for a new tree.",
                "operationId": "4211f0fc512d4ae4547f7ead6fbbfed3",
                "parameters": [
                    {
                        "name": "X-Header-Timezone",
                        "in": "header",
                        "description": "Get date fields in specific Timezone",
                        "required": false,
                        "examples": {
                            "Caracas": {
                                "summary": "Caracas",
                                "value": "America/Caracas"
                            },
                            "Bogota": {
                                "summary": "Bogota",
                                "value": "America/Bogota"
                            },
                            "Los Angeles": {
                                "summary": "Los Angeles",
                                "value": "America/Los_Angeles"
                            }
                        }
                    },
                    {
                        "name": "X-Header-Lang",
                        "in": "header",
                        "description": "Get title field in specific language",
                        "required": false,
                        "examples": {
                            "en": {
                                "summary": "English",
                                "value": "en"
                            },
                            "es": {
                                "summary": "Spanish",
                                "value": "es"
                            },
                            "fr": {
                                "summary": "French",
                                "value": "fr"
                            }
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "parent": {
                                        "oneOf": [
                                            {
                                                "type": "string"
                                            },
                                            {
                                                "type": "integer"
                                            }
                                        ]
                                    }
                                },
                                "type": "object"
                            },
                            "example": {
                                "parent": 1
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "New node created."
                    },
                    "404": {
                        "description": "Parent node not found."
                    },
                    "500": {
                        "description": "An error has happened."
                    }
                }
            }
        },
        "/api/nodes_with_children": {
            "get": {
                "tags": [
                    "nodes"
                ],
                "summary": "All nodes with children paginated",
                "description": "Display a listing of the resource.\nShowing all nodes with children paginated.",
                "operationId": "71c76059d699bc3ae4a14ecf9f0de74a",
                "parameters": [
                    {
                        "name": "X-Header-Timezone",
                        "in": "header",
                        "description": "Get date fields in specific Timezone",
                        "required": false,
                        "examples": {
                            "Caracas": {
                                "summary": "Caracas",
                                "value": "America/Caracas"
                            },
                            "Bogota": {
                                "summary": "Bogota",
                                "value": "America/Bogota"
                            },
                            "Los Angeles": {
                                "summary": "Los Angeles",
                                "value": "America/Los_Angeles"
                            }
                        }
                    },
                    {
                        "name": "X-Header-Lang",
                        "in": "header",
                        "description": "Get title field in specific language",
                        "required": false,
                        "examples": {
                            "en": {
                                "summary": "English",
                                "value": "en"
                            },
                            "es": {
                                "summary": "Spanish",
                                "value": "es"
                            },
                            "fr": {
                                "summary": "French",
                                "value": "fr"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "All nodes with children paginated."
                    },
                    "500": {
                        "description": "An error has happened."
                    }
                }
            }
        },
        "/api/nodes_without_children": {
            "get": {
                "tags": [
                    "nodes"
                ],
                "summary": "All nodes without children paginated",
                "description": "Display a listing of the resource.\nShowing all nodes without children paginated.",
                "operationId": "b207385cec47e1113c9515e57dd489db",
                "parameters": [
                    {
                        "name": "X-Header-Timezone",
                        "in": "header",
                        "description": "Get date fields in specific Timezone",
                        "required": false,
                        "examples": {
                            "Caracas": {
                                "summary": "Caracas",
                                "value": "America/Caracas"
                            },
                            "Bogota": {
                                "summary": "Bogota",
                                "value": "America/Bogota"
                            },
                            "Los Angeles": {
                                "summary": "Los Angeles",
                                "value": "America/Los_Angeles"
                            }
                        }
                    },
                    {
                        "name": "X-Header-Lang",
                        "in": "header",
                        "description": "Get title field in specific language",
                        "required": false,
                        "examples": {
                            "en": {
                                "summary": "English",
                                "value": "en"
                            },
                            "es": {
                                "summary": "Spanish",
                                "value": "es"
                            },
                            "fr": {
                                "summary": "French",
                                "value": "fr"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "All nodes without children paginated."
                    },
                    "500": {
                        "description": "An error has happened."
                    }
                }
            }
        },
        "/api/nodes/{id}": {
            "get": {
                "tags": [
                    "nodes"
                ],
                "summary": "Showing a node",
                "description": "Display the specified resource.\nShow a node in the storage.",
                "operationId": "44ba01270997ee7e6c90e9276433d4b3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id to search the node in storage",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "int": {
                                "summary": "An integer value.",
                                "value": "1"
                            }
                        }
                    },
                    {
                        "name": "X-Header-Timezone",
                        "in": "header",
                        "description": "Get date fields in specific Timezone",
                        "required": false,
                        "examples": {
                            "Caracas": {
                                "summary": "Caracas",
                                "value": "America/Caracas"
                            },
                            "Bogota": {
                                "summary": "Bogota",
                                "value": "America/Bogota"
                            },
                            "Los Angeles": {
                                "summary": "Los Angeles",
                                "value": "America/Los_Angeles"
                            }
                        }
                    },
                    {
                        "name": "X-Header-Lang",
                        "in": "header",
                        "description": "Get title field in specific language",
                        "required": false,
                        "examples": {
                            "en": {
                                "summary": "English",
                                "value": "en"
                            },
                            "es": {
                                "summary": "Spanish",
                                "value": "es"
                            },
                            "fr": {
                                "summary": "French",
                                "value": "fr"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Node information."
                    },
                    "404": {
                        "description": "Node not found."
                    },
                    "500": {
                        "description": "An error has happened."
                    }
                }
            },
            "delete": {
                "tags": [
                    "nodes"
                ],
                "summary": "Remove a node",
                "description": "Remove the specified resource from storage.\nDelete a node in the storage.",
                "operationId": "ec21ebe288c34a4302229c6bc806f6e0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id to search and delete the node in storage",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "int": {
                                "summary": "An integer value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted: No content."
                    },
                    "403": {
                        "description": "Error: Forbidden."
                    },
                    "404": {
                        "description": "Node not found."
                    },
                    "500": {
                        "description": "An error has happened."
                    }
                }
            },
            "patch": {
                "tags": [
                    "nodes"
                ],
                "summary": "Updating a node",
                "description": "Update the specified resource in storage.\nUpdate a node information.",
                "operationId": "004b21cf15a9387f6cca9eca02ce3b83",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id to search and update the node in storage",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "int": {
                                "summary": "An integer value.",
                                "value": "1"
                            }
                        }
                    },
                    {
                        "name": "X-Header-Timezone",
                        "in": "header",
                        "description": "Get date fields in specific Timezone",
                        "required": false,
                        "examples": {
                            "Caracas": {
                                "summary": "Caracas",
                                "value": "America/Caracas"
                            },
                            "Bogota": {
                                "summary": "Bogota",
                                "value": "America/Bogota"
                            },
                            "Los Angeles": {
                                "summary": "Los Angeles",
                                "value": "America/Los_Angeles"
                            }
                        }
                    },
                    {
                        "name": "X-Header-Lang",
                        "in": "header",
                        "description": "Get title field in specific language",
                        "required": false,
                        "examples": {
                            "en": {
                                "summary": "English",
                                "value": "en"
                            },
                            "es": {
                                "summary": "Spanish",
                                "value": "es"
                            },
                            "fr": {
                                "summary": "French",
                                "value": "fr"
                            }
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "parent": {
                                        "oneOf": [
                                            {
                                                "type": "string"
                                            },
                                            {
                                                "type": "integer"
                                            }
                                        ]
                                    },
                                    "title": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            },
                            "example": {
                                "parent": 1,
                                "title": "un dos tres"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Node information updated."
                    },
                    "403": {
                        "description": "Error: Forbidden. Bad Allocation"
                    },
                    "404": {
                        "description": "Node not found."
                    },
                    "500": {
                        "description": "An error has happened."
                    }
                }
            }
        }
    }
}